requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
openai==1.12.0